moved nav.js to app.js for consistancy
