+++ simulator/frontend/src/App.js
@@ -1,21 +1,104 @@
 import React from 'react';
-import { HashRouter as Router, Route, Redirect } from 'react-router-dom';
-import CssBaseline from '@material-ui/core/CssBaseline';
+import {
+  ThemeProvider,
+  createMuiTheme,
+  makeStyles,
+  AppBar,
+  Toolbar,
+  Typography,
+  Button,
+  IconButton,
+} from '@material-ui/core';
+import MenuIcon from '@material-ui/icons/Menu';
+import {
+  BrowserRouter as Router, Switch, Route, Link,
+} from 'react-router-dom';
 import Home from './pages/home';
 import SimulationWindow from './pages/simulationWindow';
-import './App.css';
 
-export default function App() {
+const useStyles = makeStyles((theme) => ({
+  root: {
+    flexGrow: 1,
+    overflowX: 'hidden',
+    overflowY: 'hidden',
+  },
+  menuButton: {
+    marginRight: theme.spacing(2),
+    backgroundColor: '#ced4da',
+  },
+  title: {
+    flexGrow: 1,
+    color: '#FFF',
+  },
+  link: {
+    '&:hover': {
+      color: '#000000',
+      textDecoration: 'none',
+    },
+  },
+}));
+
+const theme = createMuiTheme({
+  palette: {
+    primary: {
+      main: '#881c1c',
+    },
+  },
+  background: {
+    color: '#444e58',
+  },
+});
+
+export const ScenariosContext = React.createContext();
+
+export default function Nav() {
+  const classes = useStyles();
+
   return (
-    <Router>
-      <Route exact path="/">
-        <Redirect to="/simulator" />
-      </Route>
-      <Route path="/simulator" component={Home} />
-      <Route
-        path="/simulation/:id"
-        render={(props) => <SimulationWindow {...props} />}
-      />
-    </Router>
+    <div className={classes.root}>
+      <Router>
+        <div>
+          <ThemeProvider theme={theme}>
+            <AppBar position="static" color="primary">
+              <Toolbar>
+                <IconButton
+                  edge="start"
+                  className={classes.menuButton}
+                  color="inherit"
+                  aria-label="menu"
+                >
+                  <MenuIcon />
+                </IconButton>
+                <Typography variant="h6" className={classes.title}>
+                  <Link className={classes.link} to="/">
+                    <Button className={classes.title} color="inherit">
+                      Home
+                    </Button>
+                  </Link>
+                </Typography>
+                <Button
+                  onClick={() => (window.location.href = '/Shibboleth.sso/Logout?return=/')}
+                  color="inherit"
+                >
+                  LogOut
+                </Button>
+              </Toolbar>
+            </AppBar>
+
+            {/* A <Switch> looks through its children <Route>s and
+            renders the first one that matches the current URL. */}
+            <Switch>
+              <Route path="/" exact>
+                <Home />
+              </Route>
+              <Route
+                path="/simulation/:id"
+                render={(props) => <SimulationWindow {...props} />}
+              />
+            </Switch>
+          </ThemeProvider>
+        </div>
+      </Router>
+    </div>
   );
 }
